pyproject.toml
[project]
name = "vq-vae"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "neuraloperator>=2.0.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "vector-quantize-pytorch>=1.27.21",
]
